pyproject.toml
[project]
name="lux-discord"
version="0.0.1"
authors = [
    {name="Euxcbsks", email="hawhaw02030746@gmail.com"}
]
description = "Python Discord bot framework"
requires-python = ">=3.11"
dependencies = [
    "click",
    "disnake",
    "motor",
    "pymongo[srv]"
]

[project.optional-dependencies]
dotenv = [
    "python-dotenv"
]
colorlog = [
    "colorlog"
]

[build-system]
requires = [
    "setuptools >= 65.5.0"
]
build-backend = "setuptools.build_meta"
